restructure math functions
